Update task02.md
This is probably where I got stuck. 
I also missed the big picture of "The point of Task 2, rather, is that the deck should be cloned before returning it." by getting caught up on `spaced repetition algorithm` part and getting stuck on step 1 `unit test`. But telling it straight up might get them to not focus as much on the task at hand as a real-world exercise.